feat(add): add and restructure templates
