+++ src/utils/templates/error.ts
@@ -1,0 +1,25 @@
+import type { Template } from '.'
+
+const error: Template = () => ({
+  path: `error.vue`,
+  contents: `
+<script setup lang="ts">
+import type { NuxtError } from '#app'
+
+const props = defineProps({
+  error: Object as () => NuxtError
+})
+</script>
+
+<template>
+  <div>
+    <h1>{{ error.statusCode }}</h1>
+    <NuxtLink to="/">Go back home</NuxtLink>
+  </div>
+</template>
+
+<style scoped></style>
+`,
+})
+
+export { error }
